+++ cosmos/precompile/distribution/distribution.go
@@ -27,7 +27,6 @@ import (
 	bindings "pkg.berachain.dev/polaris/contracts/bindings/cosmos/precompile"
 	cosmlib "pkg.berachain.dev/polaris/cosmos/lib"
 	"pkg.berachain.dev/polaris/cosmos/precompile"
-	"pkg.berachain.dev/polaris/eth/accounts/abi"
 	ethprecompile "pkg.berachain.dev/polaris/eth/core/precompile"
 )
 
@@ -38,13 +37,10 @@ type Contract struct {
 
 // NewPrecompileContract returns a new instance of the bank precompile contract.
 func NewPrecompileContract() ethprecompile.StatefulImpl {
-	var contractAbi abi.ABI
-	if err := contractAbi.UnmarshalJSON([]byte(bindings.DistributionModuleMetaData.ABI)); err != nil {
-		panic(err)
-	}
 	return &Contract{
 		BaseContract: precompile.NewBaseContract(
-			contractAbi, cosmlib.AccAddressToEthAddress(
-				authtypes.NewModuleAddress(distributiontypes.ModuleName))),
+			bindings.DistributionModuleMetaData.ABI,
+			cosmlib.AccAddressToEthAddress(authtypes.NewModuleAddress(distributiontypes.ModuleName)),
+		),
 	}
 }
