+++ docs/web/pages/docs/precompile-cosmos-overview.mdx
@@ -3,15 +3,15 @@
 ## Stateful Precompile information
 A Stateful Precompile is a precompiled contract that can alter chain state.
 
-| Precompile Contract                                           | Contract Address                                 | Solidity Interface                                               | Associated Cosmos Module                          |
-| :------------------------------------------------------------ | :----------------------------------------------- | :------------------------------------------------ | :-------------------------------------------------| 
-| Staking Precompile                                            | `0xd9A998CaC66092748FfEc7cFBD155Aae1737C2fF`       |   [**staking.sol**](https://github.com/berachain/polaris/blob/main/precompile/contracts/solidity/src/staking.sol)       | [**Staking Module**](https://docs.cosmos.network/v0.47/modules/staking)      | 
-| Multicall Precompile                                          | Coming soon       |   [**multicall.sol**]()     | N/A                                               |
-| Bank Precompile                                               | `0x4381dC2aB14285160c808659aEe005D51255adD7`       |   [**bank.sol**](https://github.com/berachain/polaris/blob/main/precompile/contracts/solidity/src/bank.sol)          | [**Bank Module**](https://docs.cosmos.network/v0.47/modules/bank)         | 
-| IBC Precompile                                                | Coming soon       |   [**ibc.sol**]()           | [**IBC**](https://ibc.cosmos.network/)          | 
-| Governance Precompile                                         | Coming soon       |   [**governance.sol**]()    | [**Governance Module**](https://docs.cosmos.network/v0.47/modules/gov)   |
-| Address Utility Precompile                                    | `0x19`      |   [**address.sol**](https://github.com/berachain/polaris/blob/main/precompile/contracts/solidity/src/address.sol)       | N/A                                               |
-| Distribution Precompile | `0x93354845030274cD4bf1686Abd60AB28EC52e1a7` | [**distribution.sol**](https://github.com/berachain/polaris/blob/main/precompile/contracts/solidity/src/distribution.sol) | [**Distribution Module**](https://docs.cosmos.network/v0.47/modules/distribution) |
+| Precompile Contract                                           | Contract Address                                 |   Solidity Interface                                                                                                      | Associated Module                          |
+| :------------------------------------------------------------ | :----------------------------------------------- |   :------------------------------------------------                                                                       | :-------------------------------------------------| 
+| Staking Precompile                                            | `0xd9A998CaC66092748FfEc7cFBD155Aae1737C2fF`     |   [staking.sol](https://github.com/berachain/polaris/blob/main/contracts/src/cosmos/precompile/staking.sol)               | [Staking Module](https://docs.cosmos.network/v0.47/modules/staking)      | 
+| Multicall Precompile                                          | Coming soon                                      |   [multicall.sol]()                                                                                                   | [Ethereum Multicall](https://github.com/mds1/multicall)                                               |
+| Bank Precompile                                               | `0x4381dC2aB14285160c808659aEe005D51255adD7`     |   [bank.sol](https://github.com/berachain/polaris/blob/main/contracts/src/cosmos/precompile/bank.sol)                 | [Bank Module](https://docs.cosmos.network/v0.47/modules/bank)         | 
+| IBC Precompile                                                | Coming soon                                      |   [ibc.sol]()                                                                                                         | [IBC](https://ibc.cosmos.network/)                  | 
+| Governance Module Precompile                                  | `0x7b5Fe22B5446f7C62Ea27B8BD71CeF94e03f3dF2`     |   [governance.sol](https://github.com/berachain/polaris/blob/main/contracts/src/cosmos/precompile/governance.sol)     | [Governance Module](https://docs.cosmos.network/v0.47/modules/gov)   |
+| Auth Module Precompile                                        | `0xBDF49C3C3882102fc017FFb661108c63a836D065`     |   [address.sol](https://github.com/berachain/polaris/blob/main/contracts/src/cosmos/precompile/auth.sol)              | [Auth Module](https://docs.cosmos.network/main/modules/auth)    |
+| Distribution Module Precompile                                | `0x93354845030274cD4bf1686Abd60AB28EC52e1a7`     |   [distribution.sol](https://github.com/berachain/polaris/blob/main/contracts/src/cosmos/precompile/distribution.sol) | [Distribution Module](https://docs.cosmos.network/v0.47/modules/distribution) |
 
 ## Stateful Precompile ABI Generation
